project/tinyurl/requirements.txt
Flask
flask-restful
cassandra-driver
grpcio-tools
